+++ ft_chrindex.c
@@ -6,7 +6,7 @@
 /*   By: pholster <pholster@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/02/03 22:38:37 by pholster       #+#    #+#                */
-/*   Updated: 2019/02/03 22:38:37 by pholster      ########   odam.nl         */
+/*   Updated: 2019/02/25 13:06:54 by pholster      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_chrindex(const char *str, int c)
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
 	if (str == NULL)
